+++ src/css/colors.variables.scss
@@ -5,3 +5,7 @@ $gray-chinese-silver: #ccc7c7;
 $gray-light: #c9c9c9;
 $gray-dark: #474343;
 $gray-border: #dfdfdf;
+
+$colors: (
+  --q-gray-light: #{$gray-light},
+);
