+++ src/pages/User/components/UserData/styles.scss
@@ -1,3 +1,5 @@
+@use '../../../../css/colors.variables.scss';
+
 .UserInformation {
   background-color: $secondary;
   margin-top: 0.2rem;
