+++ src/pages/User/mock.ts
@@ -1,22 +1,4 @@
-export type Subscriptions = {
-  __typename: 'subscriptions';
-  expiration: string;
-};
-
-export type MemberInformation = {
-  __typename: 'members';
-  first_names: string;
-  last_names: string;
-  email: string;
-  subscriptions: Subscriptions[];
-};
-
-export type UserType = {
-  __typename: string;
-  avatar_url?: string;
-  member_code: string;
-  member_info: MemberInformation[];
-};
+import { UserType } from '../../utils/apollo.types';
 
 export const user: UserType = {
   __typename: 'users',
